third_party/WebKit/LayoutTests/http/tests/serviceworker/serviceworkerglobalscope-scope.html

<!DOCTYPE html>
<title>ServiceWorkerGlobalScope: scope property</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
function scope_test(name, scope) {
  var t = async_test('Verify the scope property: ' + name);
  t.step(function() {
      var expectedScope, options;
      if (scope) {
        expectedScope = new URL(scope, document.location).toString();
      } else {
        expectedScope = new URL('/', document.location).toString();
      }

      var registration;
      service_worker_unregister_and_register(
          t, 'resources/serviceworkerglobalscope-scope-worker.js', scope)
        .then(function(r) {
            registration = r;
            assert_equals(registration.waiting, null,
                          'registration.waiting should be null');
            assert_equals(registration.active, null,
                          'registration.active should be null');
            assert_equals(registration.installing, null,
                          'registration.installing should be null');
            return wait_for_update(t, registration);
          })
        .then(function(worker) {
            assert_equals(registration.waiting, null,
                          'registration.waiting should be null');
            assert_equals(registration.active, null,
                          'registration.active should be null');
            assert_equals(registration.installing, worker,
                          'registration.installing should not be null');
            return new Promise(function(resolve) {
                var messageChannel = new MessageChannel();
                messageChannel.port1.onmessage = resolve;
                worker.postMessage('', [messageChannel.port2]);
              });
          })
        .then(function(e) {
            var message = e.data;
            assert_equals(message.initialScope, expectedScope,
                          'Worker should see the scope on eval.');
            assert_equals(message.currentScope, expectedScope,
                          'Worker scope should not change.');
            service_worker_unregister_and_done(t, scope);
          })
        .catch(unreached_rejection(t));
    });
}

scope_test('default');
scope_test('relative path', '/a/b/c/');
scope_test('absolute url', 'http://127.0.0.1:8000/');

</script>
